Rename src/apps/admin/text_templates to src/apps/admin/textTemplates
