+++ app/packs/src/apps/admin/textTemplates/TextTemplateIcon.js
